(split)bug: 17731
Also, make sure that the literal size is a long to avoid rolling over an int.

http://bugzilla.zimbra.com/show_bug.cgi?id=17731

Copied from Perforce
 Change: 56595

+++ src/java/com/zimbra/cs/imap/TcpImapRequest.java
@@ -42,7 +42,7 @@ public class TcpImapRequest extends ImapRequest {
 
     private TcpServerInputStream mStream;
     private int mLiteral = -1;
-    private int mSize;
+    private long mSize;
     private boolean mUnlogged;
 
     TcpImapRequest(String line, ImapHandler handler) {
